NTP-969: Move Enquiry Builder start button to the Search Results page (#249)
* NTP-969: Initial commit with slight restructure to enquiry pages, so starts from search and has guidance and confirm pages

* NTP-969: Refactored Search and Subjects pages, so that the query and mediator code is in Application for consistency and ease of reuse
