+++ UI/Pages/Cookies.cshtml
@@ -165,7 +165,7 @@
                 <td class="govuk-table__cell">Stores tuition partners you've added to your price comparison list</td>
                 <td class="govuk-table__cell">24 hours</td>
             </tr>
-            @*TODO - confirm wording for session cookie*@
+            @*TODO - Content review*@
             <tr>
                 <td class="govuk-table__cell">.FindATuitionPartner.Session</td>
                 <td class="govuk-table__cell">Used for temporary storage of the enquiry details if used</td>
