+++ UI/Pages/Enquiry/Respond/AllEnquirerResponses.cshtml
@@ -1,5 +1,5 @@
-@page
-@model UI.Pages.EnquirerViewAllResponses
+﻿@page
+@model UI.Pages.Enquiry.Respond.AllEnquirerResponses
 @using GovUk.Frontend.AspNetCore.TagHelpers
 @{
     ViewData["Title"] = "Enquirer View Responses";
