+++ UI/Pages/Enquiry/Respond/ResponseConfirmation.cshtml
@@ -1,0 +1,27 @@
+﻿@page
+@model UI.Pages.Enquiry.Respond.ResponseConfirmation
+@using GovUk.Frontend.AspNetCore.TagHelpers
+@{
+    ViewData["Title"] = "Enquiry response confirmation";
+    ViewData["IncludePrintPage"] = true;
+}
+
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        <govuk-notification-banner type="Success">
+            <p class="govuk-notification-banner__heading">
+                Application complete
+            </p>
+        </govuk-notification-banner>
+        <p class="govuk-body">
+            We have sent you a confirmation email.
+        </p>
+        <h1 class="govuk-heading-l">What happens next</h1>
+        <p class="govuk-body">
+            We've sent your application.
+        </p>
+        <p class="govuk-body">
+            They will contact you either to confirm your registration, or to ask for more information.
+        </p>
+    </div>
+</div>
