+++ UI/Pages/Enquiry/Respond/ResponseConfirmation.cshtml.cs
@@ -1,0 +1,9 @@
+namespace UI.Pages.Enquiry.Respond
+{
+    public class ResponseConfirmation : PageModel
+    {
+        public void OnGet()
+        {
+        }
+    }
+}
