+++ UI/Pages/SearchResults.cshtml.cs
@@ -1,9 +1,6 @@
-using Application.Common.Interfaces;
+using System.Threading;
 using Application.Common.Models;
-using Domain;
-using Domain.Search;
-using FluentValidationResult = FluentValidation.Results.ValidationResult;
-using KeyStage = Domain.Enums.KeyStage;
+using Application.Validators;
 using TuitionType = Domain.Enums.TuitionType;
 
 namespace UI.Pages;
@@ -13,7 +10,7 @@ public class SearchResults : PageModel
     private readonly IMediator _mediator;
     public SearchResults(IMediator mediator) => _mediator = mediator;
 
-    public ResultsModel Data { get; set; } = new();
+    public SearchResultsModel Data { get; set; } = new();
     [BindProperty(SupportsGet = true)] public List<string> CompareListedTuitionPartners { get; set; } = new();
     public List<SelectableTuitionPartnerModel> SelectableTuitionPartners { get; set; } = new();
     public int TotalCompareListedTuitionPartners { get; set; }
@@ -21,34 +18,45 @@ public class SearchResults : PageModel
 
     // Name keys : Tps => TuitionPartners,Tp => TuitionPartner
 
-    public async Task OnGet(Query data)
+    public async Task<IActionResult> OnGet(GetSearchResultsQuery data)
     {
-        await CommonOnGetPostLogic(data);
+        return await CommonOnGetPostLogic(data);
     }
 
-    public async Task OnPost(Query data)
+    public async Task<IActionResult> OnPost(GetSearchResultsQuery data)
     {
-        await CommonOnGetPostLogic(data);
+        return await CommonOnGetPostLogic(data);
     }
 
-    public async Task OnGetClearAllFilters(string postcode)
+    public async Task<IActionResult> OnGetClearAllFilters(string postcode)
     {
-        Data = await _mediator.Send(new Query
-        { Postcode = postcode, Subjects = null, TuitionType = TuitionType.Any, KeyStages = null });
-
-        await SetSelectableTuitionPartners();
+        return await CommonOnGetPostLogic(new GetSearchResultsQuery { Postcode = postcode });
     }
 
-    private async Task CommonOnGetPostLogic(Query data)
+    private async Task<IActionResult> CommonOnGetPostLogic(GetSearchResultsQuery data)
     {
         data.TuitionType ??= TuitionType.Any;
-        if (data.KeyStages == null && data.Subjects != null)
-        {
-            data.KeyStages = Enum.GetValues(typeof(KeyStage)).Cast<KeyStage>()
-                .Where(x => string.Join(" ", data.Subjects).Contains(x.ToString())).ToArray();
-        }
+        data.KeyStages = data.KeyStages.UpdateFromSubjects(data.Subjects);
 
-        Data = await _mediator.Send(data);
+        Data = new SearchResultsModel(data)
+        {
+            AllSubjects = await _mediator.Send(new GetWhichSubjectQuery
+            {
+                Subjects = data.Subjects,
+            }),
+            AllTuitionTypes = AllTuitionTypes
+        };
+
+        ModelState.Clear();
+        MapErrors(await new SearchResultValidator().ValidateAsync(data));
+        if (!ModelState.IsValid) return Page();
+
+        var searchResultData = await _mediator.Send(data);
+        MapErrors(searchResultData.Validation);
+        if (!ModelState.IsValid) return Page();
+
+        Data.Results = searchResultData.Results;
+
         Data.From = ReferrerList.SearchResults;
 
         //Clear compare list TuitionType if has been changed on compare list
@@ -58,12 +66,17 @@ public class SearchResults : PageModel
         }
         Data.PreviousTuitionType = data.TuitionType;
 
-        if (!Data.Validation.IsValid)
-            foreach (var error in Data.Validation.Errors)
-                ModelState.AddModelError($"Data.{error.PropertyName}", error.ErrorMessage);
-
         await SetSelectableTuitionPartners();
+        return Page();
     }
+
+    private void MapErrors(ValidationResult? validationResult)
+    {
+        if (validationResult != null && !validationResult.IsValid)
+            foreach (var error in validationResult.Errors)
+                ModelState.AddModelError($"Data.{error.PropertyName}", error.ErrorMessage);
+    }
+
     private async Task SetSelectableTuitionPartners()
     {
         if (Data.Results != null)
@@ -71,7 +84,7 @@ public class SearchResults : PageModel
     }
 
     private async Task<List<SelectableTuitionPartnerModel>> GetSelectableTuitionPartners(
-        ResultsModel resultsModel, IEnumerable<string> selectedTuitionPartners)
+        SearchResultsModel resultsModel, IEnumerable<string> selectedTuitionPartners)
     {
         var selectableTuitionPartnerModels = new List<SelectableTuitionPartnerModel>();
         selectedTuitionPartners = selectedTuitionPartners.ToList();
@@ -144,193 +157,12 @@ public class SearchResults : PageModel
 
         return selectableTuitionPartnerModels;
     }
-    public record Query : SearchModel, IRequest<ResultsModel>
-    {
-        public TuitionType? PreviousTuitionType { get; set; } = null;
-    };
 
-    public record ResultsModel : SearchModel
-    {
-        public ResultsModel()
+    private static List<TuitionType> AllTuitionTypes =>
+        new()
         {
-        }
-
-        public ResultsModel(SearchModel query) : base(query)
-        {
-        }
-
-        public Dictionary<KeyStage, Selectable<string>[]> AllSubjects { get; set; } = new();
-        public IEnumerable<TuitionType> AllTuitionTypes { get; set; } = new List<TuitionType>();
-
-        public TuitionPartnersResult? Results { get; set; }
-        public FluentValidationResult Validation { get; internal set; } = new();
-        public TuitionType? PreviousTuitionType { get; set; }
-    }
-
-    private sealed class Validator : AbstractValidator<Query>
-    {
-        public Validator()
-        {
-            RuleFor(m => m.Postcode)
-                .NotEmpty()
-                .WithMessage("Enter a postcode");
-
-            RuleFor(m => m.Postcode)
-                .Matches(StringConstants.PostcodeRegExp)
-                .WithMessage("Enter a real postcode")
-                .When(m => !string.IsNullOrEmpty(m.Postcode));
-
-            RuleForEach(m => m.Subjects)
-                .Must(x => KeyStageSubject.TryParse(x, out var _));
-        }
-    }
-
-    public class Handler : IRequestHandler<Query, ResultsModel>
-    {
-        private readonly ILocationFilterService _locationService;
-        private readonly ITuitionPartnerService _tuitionPartnerService;
-        private readonly ILookupDataService _lookupDataService;
-        private readonly IMediator _mediator;
-
-        public Handler(ILocationFilterService locationService, ITuitionPartnerService tuitionPartnerService,
-            ILookupDataService lookupDataService, IMediator mediator)
-        {
-            _locationService = locationService;
-            _tuitionPartnerService = tuitionPartnerService;
-            _lookupDataService = lookupDataService;
-            _mediator = mediator;
-        }
-
-        public async Task<ResultsModel> Handle(Query request, CancellationToken cancellationToken)
-        {
-            var queryResponse = new ResultsModel(request) with
-            {
-                AllSubjects = await GetSubjectsList(request, cancellationToken),
-                AllTuitionTypes = AllTuitionTypes,
-            };
-
-            var searchResults = await GetSearchResults(request, cancellationToken);
-
-            return searchResults switch
-            {
-                SuccessResult => queryResponse with
-                {
-                    Results = searchResults.Data,
-                },
-                Domain.ValidationResult error => queryResponse with
-                {
-                    Validation = new FluentValidationResult(error.Failures)
-                },
-                ErrorResult error => queryResponse with
-                {
-                    Validation = new FluentValidationResult(new[]
-                    {
-                        new ValidationFailure("Postcode", error.ToString()),
-                    }),
-                },
-                _ => queryResponse with { Validation = UnknownError() },
-            };
-
-            static FluentValidationResult UnknownError() =>
-                new(new[] { new ValidationFailure("", "An unknown problem occurred") });
-        }
-
-        private static KeyStage[] AllKeyStages =>
-            new[]
-            {
-                KeyStage.KeyStage1,
-                KeyStage.KeyStage2,
-                KeyStage.KeyStage3,
-                KeyStage.KeyStage4,
-            };
-
-        private static List<TuitionType> AllTuitionTypes =>
-            new()
-            {
-                TuitionType.Any,
-                TuitionType.InSchool,
-                TuitionType.Online,
-            };
-
-        private async Task<Dictionary<KeyStage, Selectable<string>[]>> GetSubjectsList(Query request,
-            CancellationToken cancellationToken)
-        {
-            return await _mediator.Send(new WhichSubjects.Query
-            {
-                KeyStages = AllKeyStages,
-                Subjects = request.Subjects,
-            }, cancellationToken);
-        }
-
-        private async Task<IResult<TuitionPartnersResult>> GetSearchResults(Query request,
-            CancellationToken cancellationToken)
-        {
-            var location = await GetSearchLocation(request, cancellationToken);
-
-            if (location is IErrorResult error)
-            {
-                return error.Cast<TuitionPartnersResult>();
-            }
-
-            var results = await FindTuitionPartners(
-                location.Data,
-                request,
-                cancellationToken);
-
-            var result = new TuitionPartnersResult(results, location.Data.LocalAuthority);
-
-            return Result.Success(result);
-        }
-
-        private async Task<IResult<LocationFilterParameters>> GetSearchLocation(Query request,
-            CancellationToken cancellationToken)
-        {
-            var validationResults = await new Validator().ValidateAsync(request, cancellationToken);
-
-            return !validationResults.IsValid
-                ? Result.Invalid<LocationFilterParameters>(validationResults.Errors)
-                : (await _locationService.GetLocationFilterParametersAsync(request.Postcode!)).TryValidate();
-        }
-
-        private async Task<IEnumerable<TuitionPartnerResult>> FindTuitionPartners(
-            LocationFilterParameters parameters,
-            Query request,
-            CancellationToken cancellationToken)
-        {
-            var keyStageSubjects = request.Subjects?.ParseKeyStageSubjects() ?? Array.Empty<KeyStageSubject>();
-
-            var subjects = await _lookupDataService.GetSubjectsAsync(cancellationToken);
-
-            var subjectFilterIds = subjects
-                                    .Where(e => keyStageSubjects.Select(x => $"{x.KeyStage}-{x.Subject}".ToSeoUrl()).Contains(e.SeoUrl))
-                                    .Select(x => x.Id);
-
-            var tuitionFilterId = request.TuitionType > 0 ? (int?)request.TuitionType : null;
-
-            var tuitionPartnersIds = await _tuitionPartnerService.GetTuitionPartnersFilteredAsync(
-                new TuitionPartnersFilter
-                {
-                    LocalAuthorityDistrictId = parameters.LocalAuthorityDistrictId,
-                    SubjectIds = subjectFilterIds,
-                    TuitionTypeId = tuitionFilterId
-                }, cancellationToken);
-
-            var tuitionPartners = await _tuitionPartnerService.GetTuitionPartnersAsync(new TuitionPartnerRequest
-            {
-                TuitionPartnerIds = tuitionPartnersIds,
-                LocalAuthorityDistrictId = parameters.LocalAuthorityDistrictId,
-                Urn = parameters.Urn
-            }, cancellationToken);
-
-            tuitionPartners = _tuitionPartnerService.OrderTuitionPartners(tuitionPartners, new TuitionPartnerOrdering
-            {
-                OrderBy = TuitionPartnerOrderBy.Random,
-                Direction = OrderByDirection.Ascending,
-                RandomSeed = TuitionPartnerOrdering.RandomSeedGeneration(parameters.LocalAuthorityDistrictCode,
-                    parameters.Postcode, subjectFilterIds, tuitionFilterId)
-            });
-
-            return tuitionPartners;
-        }
-    }
+            TuitionType.Any,
+            TuitionType.InSchool,
+            TuitionType.Online,
+        };
 }
