+++ UI/Pages/WhichKeyStages.cshtml
@@ -41,6 +41,7 @@
             <input asp-for="Data.CompareListOrderByDirection" type="hidden" />
             <input asp-for="Data.CompareListTuitionType" type="hidden" />
             <input asp-for="Data.CompareListGroupSize" type="hidden" />
+            <input asp-for="Data.CompareListShowWithVAT" type="hidden" />
 
         </form>
     </div>
